+++ mod-di-converter-storage-server/src/main/java/org/folio/dao/ProfileDao.java
@@ -3,7 +3,6 @@ package org.folio.dao;
 import io.vertx.core.Future;
 
 import java.util.Optional;
-import java.util.function.Function;
 
 /**
  * Generic data access object
@@ -53,16 +52,6 @@ public interface ProfileDao<T, S> {
   Future<T> updateProfile(T profile, String tenantId);
 
   /**
-   * Updates T entity in database with row blocking
-   *
-   * @param profileId      Profile id
-   * @param profileMutator callback for change Profile entity before save
-   * @param tenantId       tenant id
-   * @return future with updated entity
-   */
-  Future<T> updateBlocking(String profileId, Function<T, Future<T>> profileMutator, String tenantId);
-
-  /**
    * Search in database profile with the same name
    *
    * @param profileName - profile name
@@ -91,10 +80,4 @@ public interface ProfileDao<T, S> {
    */
   Future<Boolean> hardDeleteProfile(String profileId, String tenantId);
 
-  /**
-   * Retrieve total profile's number
-   * @param tenantId - tenant id
-   * @return future with number of profiles in the table
-   */
-  Future<Integer> getTotalProfilesNumber(String tenantId);
 }