+++ mod-di-converter-storage-server/src/main/java/org/folio/dao/fieldprotectionsettings/MarcFieldProtectionSettingsDaoImpl.java
@@ -2,8 +2,6 @@ package org.folio.dao.fieldprotectionsettings;
 
 import io.vertx.core.Future;
 import io.vertx.core.Promise;
-import io.vertx.sqlclient.Row;
-import io.vertx.sqlclient.RowSet;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.folio.dao.PostgresClientFactory;
@@ -12,11 +10,13 @@ import org.folio.rest.jaxrs.model.MarcFieldProtectionSettingsCollection;
 import org.folio.rest.persist.Criteria.Criteria;
 import org.folio.rest.persist.Criteria.Criterion;
 import org.folio.rest.persist.cql.CQLWrapper;
+import org.folio.rest.persist.facets.FacetField;
 import org.folio.rest.persist.interfaces.Results;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
 import javax.ws.rs.NotFoundException;
+import java.util.List;
 import java.util.Optional;
 
 import static org.folio.dao.util.DaoUtil.constructCriteria;
@@ -35,40 +35,40 @@ public class MarcFieldProtectionSettingsDaoImpl implements MarcFieldProtectionSe
 
   @Override
   public Future<MarcFieldProtectionSettingsCollection> getAll(String query, int offset, int limit, String tenantId) {
-    Promise<Results<MarcFieldProtectionSetting>> promise = Promise.promise();
     try {
       String[] fieldList = {"*"};
       CQLWrapper cql = getCQLWrapper(MARC_FIELDS_PROTECTION_SETTINGS_TABLE, query, limit, offset);
-      pgClientFactory.createInstance(tenantId).get(MARC_FIELDS_PROTECTION_SETTINGS_TABLE, MarcFieldProtectionSetting.class, fieldList, cql, true, false, promise);
+      return pgClientFactory.createInstance(tenantId)
+        .get(MARC_FIELDS_PROTECTION_SETTINGS_TABLE, MarcFieldProtectionSetting.class, fieldList, cql,true,false, (List<FacetField>) null)
+        .map(results -> new MarcFieldProtectionSettingsCollection()
+          .withMarcFieldProtectionSettings(results.getResults())
+          .withTotalRecords(results.getResultInfo().getTotalRecords()))
+        .onFailure(e -> LOGGER.warn("getAll:: Error while searching for MarcFieldProtectionSettings", e));
     } catch (Exception e) {
       LOGGER.warn("getAll:: Error while searching for MarcFieldProtectionSettings", e);
-      promise.fail(e);
+      return Future.failedFuture(e);
     }
-    return promise.future().map(results -> new MarcFieldProtectionSettingsCollection()
-      .withMarcFieldProtectionSettings(results.getResults())
-      .withTotalRecords(results.getResultInfo().getTotalRecords()));
   }
 
   @Override
   public Future<Optional<MarcFieldProtectionSetting>> getById(String id, String tenantId) {
-    Promise<Results<MarcFieldProtectionSetting>> promise = Promise.promise();
+    Criteria crit = constructCriteria(ID_FIELD, id);
     try {
-      Criteria crit = constructCriteria(ID_FIELD, id);
-      pgClientFactory.createInstance(tenantId).get(MARC_FIELDS_PROTECTION_SETTINGS_TABLE, MarcFieldProtectionSetting.class, new Criterion(crit), true, false, promise);
+      return pgClientFactory.createInstance(tenantId)
+        .get(MARC_FIELDS_PROTECTION_SETTINGS_TABLE, MarcFieldProtectionSetting.class, new Criterion(crit), true, false, (List<FacetField>) null)
+        .map(Results::getResults)
+        .map(settings -> settings.isEmpty() ? Optional.<MarcFieldProtectionSetting>empty() : Optional.of(settings.getFirst()))
+        .onFailure(e -> LOGGER.warn("getById:: Error querying MarcFieldProtectionSetting by id {}", id, e));
     } catch (Exception e) {
       LOGGER.warn("getById:: Error querying MarcFieldProtectionSetting by id {}", id, e);
-      promise.fail(e);
+      return Future.failedFuture(e);
     }
-    return promise.future()
-      .map(Results::getResults)
-      .map(settings -> settings.isEmpty() ? Optional.empty() : Optional.of(settings.get(0)));
   }
 
   @Override
   public Future<String> save(MarcFieldProtectionSetting marcFieldProtectionSetting, String tenantId) {
-    Promise<String> promise = Promise.promise();
-    pgClientFactory.createInstance(tenantId).save(MARC_FIELDS_PROTECTION_SETTINGS_TABLE, marcFieldProtectionSetting.getId(), marcFieldProtectionSetting, promise);
-    return promise.future();
+    return pgClientFactory.createInstance(tenantId)
+      .save(MARC_FIELDS_PROTECTION_SETTINGS_TABLE, marcFieldProtectionSetting.getId(), marcFieldProtectionSetting);
   }
 
   @Override
@@ -97,9 +97,10 @@ public class MarcFieldProtectionSettingsDaoImpl implements MarcFieldProtectionSe
 
   @Override
   public Future<Boolean> delete(String id, String tenantId) {
-    Promise<RowSet<Row>> promise = Promise.promise();
-    pgClientFactory.createInstance(tenantId).delete(MARC_FIELDS_PROTECTION_SETTINGS_TABLE, id, promise);
-    return promise.future().map(updateResult -> updateResult.rowCount() == 1);
+    return pgClientFactory.createInstance(tenantId)
+      .delete(MARC_FIELDS_PROTECTION_SETTINGS_TABLE, id)
+      .map(updateResult -> updateResult.rowCount() == 1)
+      .onFailure(e -> LOGGER.warn("delete:: Error deleting MARC field protection setting with id {}", id, e));
   }
 }