+++ mod-di-converter-storage-server/src/main/java/org/folio/rest/impl/ModTenantAPI.java
@@ -4,7 +4,6 @@ import io.vertx.core.AsyncResult;
 import io.vertx.core.Context;
 import io.vertx.core.Future;
 import io.vertx.core.Handler;
-import io.vertx.core.Promise;
 import io.vertx.core.Vertx;
 import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang3.StringUtils;
@@ -50,6 +49,7 @@ public class ModTenantAPI extends TenantAPI {
   private static final String DEFAULT_QM_HOLDINGS_UPDATE_JOB_PROFILE = "templates/db_scripts/defaultData/default_qm_holdings_update_job_profile.sql";
   private static final String DEFAULT_QM_AUTHORITY_CREATE_JOB_PROFILE = "templates/db_scripts/defaultData/default_qm_authority_create_job_profile.sql";
   private static final String DEFAULT_ECS_INSTANCE_AND_MARC_BIB_CREATE_JOB_PROFILE = "templates/db_scripts/defaultData/default_ecs_instance_and_marc_bib_create_job_profile.sql";
+  private static final String DEFAULT_MOSAIC_EDIFACT_MAPPING_PROFILE = "templates/db_scripts/defaultData/default_mosaic_edifact_mapping_profile.sql";
   private static final String RENAME_MODULE = "templates/db_scripts/rename_module.sql";
 
   private static final String TENANT_PLACEHOLDER = "${myuniversity}";
@@ -97,6 +97,7 @@ public class ModTenantAPI extends TenantAPI {
         .compose(m -> runSqlScript(DEFAULT_QM_HOLDINGS_UPDATE_JOB_PROFILE, headers, context))
         .compose(m -> runSqlScript(DEFAULT_ECS_INSTANCE_AND_MARC_BIB_CREATE_JOB_PROFILE, headers, context))
         .compose(m -> runSqlScript(DEFAULT_QM_AUTHORITY_CREATE_JOB_PROFILE, headers, context))
+        .compose(m -> runSqlScript(DEFAULT_MOSAIC_EDIFACT_MAPPING_PROFILE, headers, context))
         .map(num));
   }
 
@@ -109,7 +110,7 @@ public class ModTenantAPI extends TenantAPI {
         return Future.succeededFuture();
       }
 
-      String sqlScript = IOUtils.toString(inputStream, StandardCharsets.UTF_8.name());
+      String sqlScript = IOUtils.toString(inputStream, StandardCharsets.UTF_8);
       if (StringUtils.isBlank(sqlScript)) {
         return Future.succeededFuture();
       }
@@ -119,10 +120,7 @@ public class ModTenantAPI extends TenantAPI {
 
       sqlScript = sqlScript.replace(TENANT_PLACEHOLDER, tenantId).replace(MODULE_PLACEHOLDER, moduleName);
 
-      Promise<List<String>> promise = Promise.promise();
-      PostgresClient.getInstance(context.owner()).runSQLFile(sqlScript, false, promise);
-
-      return promise.future();
+      return PostgresClient.getInstance(context.owner()).runSQLFile(sqlScript, false);
     } catch (IOException e) {
       LOGGER.warn("runSqlScript:: Failed to run sql script", e);
       return Future.failedFuture(e);