+++ mod-di-converter-storage-server/src/main/java/org/folio/rest/impl/util/ExceptionHelper.java
@@ -1,6 +1,5 @@
 package org.folio.rest.impl.util;
 
-import io.vertx.core.Promise;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.folio.HttpStatus;
@@ -12,6 +11,8 @@ import javax.ws.rs.BadRequestException;
 import javax.ws.rs.NotFoundException;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
+
+import java.util.concurrent.atomic.AtomicReference;
 
 import static javax.ws.rs.core.Response.Status.BAD_REQUEST;
 import static javax.ws.rs.core.Response.Status.CONFLICT;
@@ -50,19 +51,23 @@ public final class ExceptionHelper {
         .entity(throwable.getMessage())
         .build();
     }
+
     LOGGER.error("{}", throwable.getMessage(), throwable);
-    Promise<Response> validationFuture = Promise.promise();
-    ValidationHelper.handleError(throwable, validationFuture);
-    if (validationFuture.future().isComplete()) {
-      Response response = validationFuture.future().result();
-      if (response.getStatus() == INTERNAL_SERVER_ERROR.getStatusCode()) {
-        LOGGER.warn(throwable.getMessage(), throwable);
+    AtomicReference<Response> container = new AtomicReference<>();
+    ValidationHelper.handleError(throwable, validationFuture -> {
+      if (validationFuture.succeeded()) {
+        Response response = validationFuture.result();
+        container.set(response);
+        if (response.getStatus() == INTERNAL_SERVER_ERROR.getStatusCode()) {
+          LOGGER.warn(throwable.getMessage(), throwable);
+        }
+      } else {
+        container.set(Response.status(INTERNAL_SERVER_ERROR.getStatusCode())
+          .type(MediaType.TEXT_PLAIN)
+          .entity(INTERNAL_SERVER_ERROR.getReasonPhrase())
+          .build());
       }
-      return response;
-    }
-    return Response.status(INTERNAL_SERVER_ERROR.getStatusCode())
-      .type(MediaType.TEXT_PLAIN)
-      .entity(INTERNAL_SERVER_ERROR.getReasonPhrase())
-      .build();
+    });
+    return container.get();
   }
 }