+++ mod-di-converter-storage-server/src/main/resources/templates/db_scripts/associations-migration/remove_old_association_tables.sql
@@ -1,0 +1,7 @@
+DROP TABLE IF EXISTS ${myuniversity}_${mymodule}.action_to_action_profiles CASCADE;
+DROP TABLE IF EXISTS ${myuniversity}_${mymodule}.action_to_mapping_profiles CASCADE;
+DROP TABLE IF EXISTS ${myuniversity}_${mymodule}.action_to_match_profiles CASCADE;
+DROP TABLE IF EXISTS ${myuniversity}_${mymodule}.job_to_action_profiles CASCADE;
+DROP TABLE IF EXISTS ${myuniversity}_${mymodule}.job_to_match_profiles CASCADE;
+DROP TABLE IF EXISTS ${myuniversity}_${mymodule}.match_to_action_profiles CASCADE;
+DROP TABLE IF EXISTS ${myuniversity}_${mymodule}.match_to_match_profiles CASCADE;