+++ mod-di-converter-storage-server/src/main/resources/templates/db_scripts/defaultData/default_ecs_instance_and_marc_bib_create_job_profile.sql
@@ -317,27 +317,3 @@ BEGIN
       ) ON CONFLICT DO NOTHING;
 END
 $$;
-
-INSERT INTO ${myuniversity}_${mymodule}.job_to_action_profiles (id, jsonb) values
-  ('7e1b00ad-eb12-4c27-aae7-3c4b39e97e3d',
-   '{
-     "id": "7e1b00ad-eb12-4c27-aae7-3c4b39e97e3d",
-     "order": 0,
-     "triggered": false,
-     "detailWrapperId": "213aadd4-4d7b-42e2-af6b-2591e3db0c2c",
-     "masterWrapperId": "b4b102e0-7fb4-4b77-97f1-782927839eb0",
-     "detailProfileType": "ACTION_PROFILE",
-     "masterProfileType": "JOB_PROFILE"
-   }') ON CONFLICT DO NOTHING;
-
-INSERT INTO ${myuniversity}_${mymodule}.action_to_mapping_profiles (id, jsonb) values
-  ('654f9356-8a7f-49fc-b6d2-91b08df15433',
-   '{
-     "id": "654f9356-8a7f-49fc-b6d2-91b08df15433",
-     "order": 0,
-     "triggered": false,
-     "detailWrapperId": "e4f143c6-1b7a-4ed9-b0e6-d0b8d5f5ad63",
-     "masterWrapperId": "213aadd4-4d7b-42e2-af6b-2591e3db0c2c",
-     "detailProfileType": "MAPPING_PROFILE",
-     "masterProfileType": "ACTION_PROFILE"
-   }') ON CONFLICT DO NOTHING;