+++ NEWS.md
@@ -1,3 +1,12 @@
+## 2025-04-14 2.5.0
+* [MODDICONV-425](https://folio-org.atlassian.net/browse/MODDICONV-425) Create new Default Mapping profile for Mosaic invoices
+* [MODDICONV-429](https://folio-org.atlassian.net/browse/MODDICONV-429) Upgrade module to Vert.x 5.0
+* [MODDICONV-427](https://folio-org.atlassian.net/browse/MODDICONV-427) Delete old association tables
+
+## 2025-03-13 2.4.0
+* [MODDATAIMP-1102](https://folio-org.atlassian.net/browse/MODDATAIMP-1102) Empty contributor and vendor reference properties created in order when no values in incoming record
+* [MODDICONV-409](https://folio-org.atlassian.net/browse/MODDICONV-409) Update to Java 21 mod-di-converter-storage Sunflower R1 2025
+
 ## 2024-10-29 2.3.0
 * [MODDICONV-312](https://folio-org.atlassian.net/browse/MODDICONV-312) Create general associations table
 * [MODDICONV-373](https://folio-org.atlassian.net/browse/MODDICONV-373) Disallow create/update of a job profile without actions